requirements.txt
requests==2.32.5
websocket-client==1.9.0
python-dotenv==1.2.1
pandas==2.3.3
numpy==2.2.6
